*Still working on pattern 9.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@2634 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/JournalEntries.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections;
+using System.Collections.Generic;
 using System.Data;
 using System.Reflection;
 
@@ -9,19 +10,25 @@ namespace OpenDentBusiness{
 
 		///<summary>Used when displaying the splits for a transaction.</summary>
 		public static ArrayList GetForTrans(int transactionNum) {
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				return Meth.GetObject<ArrayList>(MethodBase.GetCurrentMethod(),transactionNum);
+			}
 			string command=
 				"SELECT * FROM journalentry "
 				+"WHERE TransactionNum="+POut.PInt(transactionNum);
-			JournalEntry[] List=RefreshAndFill(command);
+			List <JournalEntry> List=RefreshAndFill(command);
 			ArrayList retVal=new ArrayList();
-			for(int i=0;i<List.Length;i++) {
+			for(int i=0;i<List.Count;i++) {
 				retVal.Add(List[i]);
 			}
 			return retVal;
 		}
 
 		///<summary>Used to display a list of entries for one account.</summary>
-		public static JournalEntry[] GetForAccount(int accountNum) {
+		public static List <JournalEntry> GetForAccount(int accountNum) {
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				return Meth.GetObject<List <JournalEntry>>(MethodBase.GetCurrentMethod(),accountNum);
+			}
 			string command=
 				"SELECT * FROM journalentry "
 				+"WHERE AccountNum="+POut.PInt(accountNum)
@@ -30,7 +37,10 @@ namespace OpenDentBusiness{
 		}
 
 		///<summary>Used in reconcile window.</summary>
-		public static JournalEntry[] GetForReconcile(int accountNum,bool includeUncleared,int reconcileNum) {
+		public static List <JournalEntry> GetForReconcile(int accountNum,bool includeUncleared,int reconcileNum) {
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				return Meth.GetObject<List <JournalEntry>>(MethodBase.GetCurrentMethod(),accountNum,includeUncleared,reconcileNum);
+			}
 			string command=
 				"SELECT * FROM journalentry "
 				+"WHERE AccountNum="+POut.PInt(accountNum)
@@ -45,11 +55,11 @@ namespace OpenDentBusiness{
 			return RefreshAndFill(command);
 		}
 
-		private static JournalEntry[] RefreshAndFill(string command) {
+		private static List <JournalEntry> RefreshAndFill(string command) {
 			DataTable table=Db.GetTable(command);
-			JournalEntry[] List=new JournalEntry[table.Rows.Count];
-			for(int i=0;i<List.Length;i++) {
-				List[i]=new JournalEntry();
+			List <JournalEntry> List=new List <JournalEntry> ();
+			for(int i=0;i<table.Rows.Count;i++) {
+				List.Add(new JournalEntry());
 				List[i].JournalEntryNum= PIn.PInt(table.Rows[i][0].ToString());
 				List[i].TransactionNum = PIn.PInt(table.Rows[i][1].ToString());
 				List[i].AccountNum     = PIn.PInt(table.Rows[i][2].ToString());
@@ -66,6 +76,10 @@ namespace OpenDentBusiness{
 
 		///<summary></summary>
 		public static void Insert(JournalEntry je) {
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				Meth.GetVoid(MethodBase.GetCurrentMethod(),je);
+				return;
+			}
 			if(je.DebitAmt<0 || je.CreditAmt<0){
 				throw new ApplicationException(Lan.g("JournalEntries","Error. Credit and debit must both be positive."));
 			}
@@ -101,6 +115,10 @@ namespace OpenDentBusiness{
 
 		///<summary></summary>
 		public static void Update(JournalEntry je) {
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				Meth.GetVoid(MethodBase.GetCurrentMethod(),je);
+				return;
+			}
 			if(je.DebitAmt<0 || je.CreditAmt<0) {
 				throw new ApplicationException(Lan.g("JournalEntries","Error. Credit and debit must both be positive."));
 			}
@@ -120,12 +138,17 @@ namespace OpenDentBusiness{
 
 		///<summary></summary>
 		public static void Delete(JournalEntry je) {
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				Meth.GetVoid(MethodBase.GetCurrentMethod(),je);
+				return;
+			}
 			string command= "DELETE FROM journalentry WHERE JournalEntryNum = "+POut.PInt(je.JournalEntryNum);
 			Db.NonQ(command);
 		}
 
 		///<summary>Used in FormTransactionEdit to synch database with changes user made to the journalEntry list for a transaction.  Must supply an old list for comparison.  Only the differences are saved.  Surround with try/catch, because it will thrown an exception if any entries are negative.</summary>
 		public static void UpdateList(ArrayList oldJournalList,ArrayList newJournalList) {
+			//No need to check RemotingRole; no call to db.
 			for(int i=0;i<newJournalList.Count;i++){
 				if(((JournalEntry)newJournalList[i]).DebitAmt<0 || ((JournalEntry)newJournalList[i]).CreditAmt<0){
 					throw new ApplicationException(Lan.g("JournalEntries","Error. Credit and debit must both be positive."));
@@ -174,6 +197,7 @@ namespace OpenDentBusiness{
 
 		///<summary>Called from FormTransactionEdit.</summary>
 		public static bool AttachedToReconcile(ArrayList journalList){
+			//No need to check RemotingRole; no call to db.
 			for(int i=0;i<journalList.Count;i++){
 				if(((JournalEntry)journalList[i]).ReconcileNum!=0){
 					return true;
@@ -184,6 +208,7 @@ namespace OpenDentBusiness{
 
 		///<summary>Called from FormTransactionEdit.</summary>
 		public static DateTime GetReconcileDate(ArrayList journalList) {
+			//No need to check RemotingRole; no call to db.
 			for(int i=0;i<journalList.Count;i++) {
 				if(((JournalEntry)journalList[i]).ReconcileNum!=0) {
 					return Reconciles.GetOne(((JournalEntry)journalList[i]).ReconcileNum).DateReconcile;
@@ -193,10 +218,14 @@ namespace OpenDentBusiness{
 		}
 
 		///<summary>Called once from FormReconcileEdit when closing.  Saves the reconcileNum for every item in the list.</summary>
-		public static void SaveList(JournalEntry[] journalList,int reconcileNum) {
+		public static void SaveList(List <JournalEntry> journalList,int reconcileNum) {
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				Meth.GetVoid(MethodBase.GetCurrentMethod(),journalList,reconcileNum);
+				return;
+			}
 			string command="UPDATE journalentry SET ReconcileNum=0 WHERE";
 			string str="";
-			for(int i=0;i<journalList.Length;i++){
+			for(int i=0;i<journalList.Count;i++){
 				if(journalList[i].ReconcileNum==0){
 					if(str!=""){
 						str+=" OR";
@@ -210,7 +239,7 @@ namespace OpenDentBusiness{
 			}
 			command="UPDATE journalentry SET ReconcileNum="+POut.PInt(reconcileNum)+" WHERE";
 			str="";
-			for(int i=0;i<journalList.Length;i++) {
+			for(int i=0;i<journalList.Count;i++) {
 				if(journalList[i].ReconcileNum==reconcileNum) {
 					if(str!="") {
 						str+=" OR";
